+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Лог изменений
 
-## 0.5.0
+## 0.5.1
 
 - рефактор LocaleContext
 - рефактор RouterContext